Shift and refactor util files
