+++ src/visualizer/sortingvisualizer/component/radixsortboxes/styles.css
@@ -1,11 +1,3 @@
-/*.radix-container {*/
-/*  align-self: flex-start;*/
-/*  box-sizing: border-box;*/
-/*  display: flex;*/
-/*  flex: 1;*/
-/*  flex-direction: column;*/
-/*}*/
-
 .data-arr {
   align-self: center;
   display: flex;
@@ -17,7 +9,7 @@
   display: flex;
   flex-direction: row;
   justify-content: center;
-  margin-top: 18%;
+  margin-top: 250px;
 }
 
 .box {
