+++ src/visualizer/sortingvisualizer/util/CountingSortUtil.js
@@ -1,0 +1,12 @@
+// Fixed array for counting sort
+export const buckets = [
+  { height: 1, count: 0 },
+  { height: 2, count: 0 },
+  { height: 3, count: 0 },
+  { height: 4, count: 0 },
+  { height: 5, count: 0 },
+  { height: 6, count: 0 },
+  { height: 7, count: 0 },
+  { height: 8, count: 0 },
+  { height: 9, count: 0 },
+];
